refactored logo into its own component and imported where needed. moved css modules into same folder as components
