+++ src/components/ui/Button.tsx
@@ -1,5 +1,5 @@
 import { FC, MouseEvent } from 'react';
-import styles from '../../styles/Button.module.css';
+import styles from './Button.module.css';
 
 type Props = {
   text: string;
