+++ src/components/ui/Logo.module.css
@@ -1,0 +1,5 @@
+.logo {
+  height: auto;
+  align-self: center;
+  margin-bottom: 15px;
+}
