+++ src/components/ui/Logo.tsx
@@ -1,0 +1,20 @@
+import { FC } from 'react';
+import logo from '../../assets/shell.png';
+import styles from './Logo.module.css';
+
+type Props = {
+  addedStyles?: { [key: string]: string };
+};
+
+const Logo: FC<Props> = ({ addedStyles }): JSX.Element => {
+  return (
+    <img
+      src={logo}
+      style={addedStyles}
+      className={styles.logo}
+      alt='Natalius logo'
+    />
+  );
+};
+
+export default Logo;
